refactor: move DateInput to it's own file
